Monster module can now return connection type. Renamed *.model.ts into *.type.ts
