+++ src/schema/ailments/ailment.type.ts
