+++ src/schema/locations/camp.type.ts
