+++ src/schema/locations/location.type.ts
