+++ src/schema/monsters/monster.type.ts
