+++ src/schema/monsters/resistance.type.ts
