+++ src/schema/monsters/weakness.type.ts
