+++ src/schema/rewards/condition.type.ts
