+++ src/schema/shared/pageResponse.type.ts
