+++ src/schema/skills/skill-rank.type.ts
