+++ src/schema/skills/skill.type.ts
